requirements.txt
python-dotenv
pandas
numpy
scikit-learn
pymongo[srv]==3.12
pymongo
certifi

# -e . 

# -e . : This is a local package and its function is to allow for
# editable installs. It refers to setup.py file
